add app user for user views & model + edit test
