+++ product/models.py
@@ -3,16 +3,16 @@ from typing import Union, Tuple
 
 from django.contrib.auth.models import User
 from django.db import models
-
+from user.models import Profile
 from tools import logger
 
 
-class Profile(models.Model):
-    """
-    models profile for add user image
-    """
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    image = models.BinaryField(null=True)
+# class Profile(models.Model):
+#     """
+#     models profile for add user image
+#     """
+#     user = models.OneToOneField(User, on_delete=models.CASCADE)
+#     image = models.BinaryField(null=True)
 
 
 class Category(models.Model):
