+++ product/urls.py
@@ -11,9 +11,9 @@ urlpatterns = [
     path('products/save/', views.SaveProduct.as_view(), name="product_save"),
     path('products/results/', views.Result.as_view(extra_context={'title': 'results'}), name="results"),
 
-    path('accounts/login/', views.LoginView.as_view(), name="login"),
-    path('accounts/logout/', views.LogoutView.as_view(), name="logout"),
-    path('accounts/dashboard/', views.DashboardUser.as_view(), name="dashboardUser"),
-    path('accounts/register/', views.RegisterUser.as_view(), name="register"),
-    path('accounts/products/', views.MyProducts.as_view(), name="myproduct"),
+    # path('accounts/login/', views.LoginView.as_view(), name="login"),
+    # path('accounts/logout/', views.LogoutView.as_view(), name="logout"),
+    # path('accounts/dashboard/', views.DashboardUser.as_view(), name="dashboardUser"),
+    # path('accounts/register/', views.RegisterUser.as_view(), name="register"),
+    # path('accounts/products/', views.MyProducts.as_view(), name="myproduct"),
 ]
