+++ requirements.txt
@@ -7,6 +7,7 @@ django-extensions==3.1.1
 fontawesome-free==5.15.3
 gunicorn==20.1.0
 idna==2.10
+Pillow==8.2.0
 psycopg2-binary==2.8.6
 python-dateutil==2.8.1
 pytz==2021.1
