+++ user/templates/user/register.html
@@ -24,9 +24,7 @@
                 {{ form.first_name }}
                 {{ form.last_name }}
                 {{ form.email }}
-                <label class="image_form fadeIn third"> Choisir une image
-                    {{ form.image }}
-                </label>
+                {{ form.image }}
                 {{ form.password1 }}
                 {{ form.password2 }}
                 <div>
